Use cp -R instead of cp -r on systems that are picky.
On most systems, these are one and the same. However, the POSIX standard
declares -r to be obsolete, and on OS X the manpage says that it may not
correctly handle special files, symbolic links, or FIFOs.

The command is used in build.sh to create and populate the build/
directory.