Fixed some configuration issues.
Lua assumes that all Linux systems support readline; this may not be
the case, so we now manually manage defines instead of relying on
the default Linux configuration in luaconf.h.

Also improved readline detection by ensuring that #include files can
be processed properly. This requires preceding them with an include
of stdio.h